+++ Sourcery/Sourcery.swift
@@ -95,6 +95,10 @@ public class Sourcery {
                             paths.append(file)
                             modules.append(target.module)
                         }
+                        for framework in target.xcframeworks {
+                            paths.append(framework.swiftInterfacePath)
+                            modules.append(target.module)
+                        }
                     }
                 }
                 result = try self.parse(from: paths, forceParse: forceParse, parseDocumentation: parseDocumentation, modules: modules, requiresFileParserCopy: hasSwiftTemplates)